% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hubverse-package.R
\docType{package}
\name{hubverse-package}
\alias{hubverse}
\alias{hubverse-package}
\title{hubverse: Easily Install and Load the 'hubverse'}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A collection of packages that enables collaborative modeling exercises through a unified framework for aggregating, visualizing, and evaluating forecasts. This package is designed to make it easy to install and load multiple 'hubverse' packages in a single step. Learn more at \url{https://hubverse.io/en/latest/index.html}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hubverse-org/hubverse}
  \item Report bugs at \url{https://github.com/hubverse-org/hubverse/issues}
}

}
\author{
\strong{Maintainer}: Li Shandross \email{lshandross@umass.edu} (\href{https://orcid.org/0009-0008-1348-1954}{ORCID})

Other contributors:
\itemize{
  \item  Consortium of Infectious Disease Modeling Hubs [copyright holder]
}

}
\keyword{internal}
